third_party/WebKit/LayoutTests/http/tests/serviceworker/postmessage.html

<!DOCTYPE html>
<title>Service Worker: postMessage</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
async_test(function(t) {
    var scope = 'resources/blank.html';
    service_worker_unregister_and_register(
        t, 'resources/postmessage-worker.js', scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(function(sw) {
          var messageChannel = new MessageChannel();
          messageChannel.port1.onmessage = t.step_func(onMessage);
          sw.postMessage({port: messageChannel.port2}, [messageChannel.port2]);
          sw.postMessage({value: 1});
          sw.postMessage({value: 2});
          sw.postMessage({done: true});
        })
      .catch(unreached_rejection(t));

    var result = [];
    var expected = [
      'Acking value: 1',
      'Acking value: 2',
    ];

    function onMessage(e) {
      var message = e.data;
      if (message === 'quit') {
        assert_array_equals(result, expected,
                            'Worker should post back expected values.');
        service_worker_unregister_and_done(t, scope);
      } else {
        result.push(message);
      }
    };
  }, 'postMessage to a ServiceWorker (and back via MessagePort)');
</script>
